third_party/blink/web_tests/platform/generic/css3/calc/padding-expected.txt

This element should have an overall padding of 25 pixels. => PASS

This element should have a left padding of 25 pixels. => PASS

This element should have a right padding of 25 pixels. => PASS

This element should have a top padding of 25 pixels. => PASS

This element should have a bottom padding of 25 pixels. => PASS

This element should have an overall padding of 25 pixels (10% of parent width of 300px minus 5px). => PASS

This element should have a left padding of 25 pixels (10% of parent width of 300px minus 5px). => PASS

This element should have a right padding of 25 pixels (10% of parent width of 300px minus 5px). => PASS

This element should have a top padding of 25 pixels (10% of parent width of 300px minus 5px). => PASS

This element should have a bottom padding of 25 pixels (10% of parent width of 300px minus 5px). => PASS
